+++ src/errors/errors-base-class/error-base-class.ts
