+++ src/errors/not-authorized-error.ts
@@ -1,4 +1,4 @@
-import { CustomError } from "./custom-error";
+import { CustomError } from "./errors-base-class/error-base-class";
 
 export class NotAuthorizedError extends CustomError {
   statusCode = 401;
