+++ src/errors/request-validation-error.ts
@@ -1,22 +1,19 @@
-import { CustomError } from "./custom-error";
+import { CustomError } from "./errors-base-class/error-base-class";
 import { ValidationError } from "express-validator";
 
 export class RequestValidationError extends CustomError {
-
   statusCode = 400;
-  
+
   constructor(public errors: ValidationError[]) {
-    
     // Calling the parent class (Error) constructor as this is a constructor of sub-class
-    super('Invalid Login Parameters.');
+    super("Invalid Login Parameters.");
 
     // Since we are extending a built-in class (Error) the following line of cade has to be added
     Object.setPrototypeOf(this, RequestValidationError.prototype);
-    
   }
 
   serializeErrors() {
-    return this.errors.map( (error) => {
+    return this.errors.map((error) => {
       if (error.type === "field") {
         return { message: error.msg, field: error.path };
       }
@@ -24,5 +21,4 @@ export class RequestValidationError extends CustomError {
       return { message: error.msg };
     });
   }
-
 }
